requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
pymupdf==1.25.5
openai==1.81.0
langchain==0.2.0
faiss-cpu==1.11.0
sentence-transformers==4.1.0
python-multipart==0.0.20